projects/Fully_operational_MVP/backend/requirements.txt
langchain-openai
langgraph
tavily-python
python-dotenv
requests
tweepy
pillow
openai
langsmith
google-generativeai
exa_py
crawl4ai
fastapi
uvicorn[standard]
